Probably final commit.
I learned a lot from this, but just figured out that this template is the minimal starter - which caused lots of confusion when looking over the Gatsby documentation, and also probably prompted need for gatsby-plugin-layout (which was a very frustrating problem).

Will probably continue to learn React within Next.js, create-react-app, or Gatsby's default starter in the future.